Working changes
- In process of:
  - Removing cJSON in favor of direct python dict usage
  - Stop using cython (it's not you, i just wanna know this C API stuff)

+++ tf2_dem_py/flags/flags.h
@@ -8,8 +8,8 @@ extern "C" {
 #endif
 
 typedef struct FLAGS_s {
-	uint16_t CHAT;
-	uint16_t GAME_EVENTS;
+	uint32_t CHAT;
+	uint32_t GAME_EVENTS;
 } FLAGS_s;
 
 extern const FLAGS_s FLAGS;
